api/requirements.txt
# Runtime
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.0
pydantic-settings==2.3.0
httpx==0.27.0

# Tests
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0
